# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/contracts": {
      "post": {
        "operationId": "ContractController_create",
        "summary": "Create new contract",
        "description": "Creates a new contract between a client and a service provider",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Contract creation data including service, client, provider, and time slots",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContractDto"
              },
              "examples": {
                "example1": {
                  "summary": "Basic contract creation",
                  "description": "Create a contract for a cleaning service",
                  "value": {
                    "serviceId": 1,
                    "clientId": 2,
                    "providerId": 3,
                    "amount": 150,
                    "clientEmail": "client@example.com",
                    "serviceName": "House Cleaning Service",
                    "timeSlots": [
                      {
                        "date": "2024-01-15",
                        "startTime": "09:00",
                        "endTime": "12:00"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contract created successfully with payment link",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contract": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "amount": {
                          "type": "number",
                          "example": 150
                        },
                        "status": {
                          "type": "string",
                          "example": "PENDING"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2024-01-10T10:00:00Z"
                        }
                      }
                    },
                    "bookings": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "date": {
                            "type": "string",
                            "example": "2024-01-15"
                          },
                          "startTime": {
                            "type": "string",
                            "example": "09:00"
                          },
                          "endTime": {
                            "type": "string",
                            "example": "12:00"
                          },
                          "status": {
                            "type": "string",
                            "example": "PENDING"
                          }
                        }
                      }
                    },
                    "paymentUrl": {
                      "type": "string",
                      "description": "Stripe payment link URL",
                      "example": "https://checkout.stripe.com/pay/cs_test_a1B2c3D4e5F6g7H8i9J0k1L2m3N4o5P6q7R8s9T0u1V2w3X4y5Z6"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or service not available",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Service not available for requested time slots"
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Service, client, or provider not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Service not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Contracts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ContractController_findAll",
        "summary": "Get all contracts",
        "description": "Retrieves all contracts in the system (admin only)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All contracts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "amount": {
                        "type": "number",
                        "example": 150
                      },
                      "status": {
                        "type": "string",
                        "example": "PAID"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2024-01-10T10:00:00Z"
                      },
                      "service": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "title": {
                            "type": "string",
                            "example": "House Cleaning Service"
                          }
                        }
                      },
                      "client": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 2
                          },
                          "firstName": {
                            "type": "string",
                            "example": "John"
                          },
                          "lastName": {
                            "type": "string",
                            "example": "Doe"
                          }
                        }
                      },
                      "provider": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 3
                          },
                          "firstName": {
                            "type": "string",
                            "example": "Jane"
                          },
                          "lastName": {
                            "type": "string",
                            "example": "Smith"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Contracts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/contracts/my-contracts": {
      "get": {
        "operationId": "ContractController_getMyContracts",
        "summary": "Get user contracts",
        "description": "Retrieves all contracts where the authenticated user is either client or provider",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User contracts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clientContracts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "amount": {
                            "type": "number",
                            "example": 150
                          },
                          "status": {
                            "type": "string",
                            "example": "PAID"
                          },
                          "timeSlots": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "date": {
                                  "type": "string",
                                  "example": "2024-01-15"
                                },
                                "startTime": {
                                  "type": "string",
                                  "example": "09:00"
                                },
                                "endTime": {
                                  "type": "string",
                                  "example": "12:00"
                                },
                                "status": {
                                  "type": "string",
                                  "example": "CONFIRMED"
                                }
                              }
                            }
                          },
                          "canReview": {
                            "type": "boolean",
                            "example": true
                          }
                        }
                      },
                      "description": "Contracts where user is the client"
                    },
                    "providerContracts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 2
                          },
                          "amount": {
                            "type": "number",
                            "example": 200
                          },
                          "status": {
                            "type": "string",
                            "example": "PAID"
                          },
                          "timeSlots": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "date": {
                                  "type": "string",
                                  "example": "2024-01-16"
                                },
                                "startTime": {
                                  "type": "string",
                                  "example": "14:00"
                                },
                                "endTime": {
                                  "type": "string",
                                  "example": "17:00"
                                },
                                "status": {
                                  "type": "string",
                                  "example": "CONFIRMED"
                                }
                              }
                            }
                          },
                          "canReview": {
                            "type": "boolean",
                            "example": false
                          }
                        }
                      },
                      "description": "Contracts where user is the provider"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Contracts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Manexo API",
    "description": "API documentation for Manexo application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateContractDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}